split compiler ops
