+++ cli/ops/mod.rs
@@ -21,6 +21,7 @@ pub mod process;
 pub mod random;
 pub mod repl;
 pub mod resources;
+pub mod runtime_compiler;
 pub mod timers;
 pub mod tls;
 pub mod web_worker;
